New Functions
Se agregan optiones por peticion fetch
Se crea funcion que reciba datos y lo envie a un callback
