+++ index.html
@@ -28,11 +28,9 @@
                     </div>
                     <div class="section-home__search container">
                         <div class="options">
-                            <select name="coin-select" id="country">
+                            <select name="coin-select" id="coins">
                                 <option value="" selected="selected">Selecciona una moneda</option>
-                                <option value="Bitcoin">Bitcoin</option>
-                                <option value="Ethereum">Ethereum</option>
-                                <option value="Tether">Tether</option>
+                               
                             </select>
                         </div>
                         <button type="submit" class="btn btn-submit--1">
