+++ src/js/main.js
@@ -15,5 +15,5 @@ window.onload = async () => {
   drawCards(await getList(url));
   printOptions(await getList(url));
   typed(await getList(url));
-  new SlimSelect({ select: "#country" });
+  new SlimSelect({ select: "#coins" });
 };
